feat: stories 6.x/7.x 후속 작업 — admin 사용자관리 + 후속 마이그레이션
- Alembic 0020/0021/0022 (검색 인덱스, 차단 컬럼, free 지연 override)
- Alembic 0024_pre_block_status — pre-block 상태 추가 (6.x 후속)
- admin/users: 사용자 검색/수정/차단 라우터 + 서비스 + 스키마
- admin/auth, deps/auth: 권한 체크 보강
- 사용자 메일 변경/차단 서비스 user_service +314
- celery_app: anomaly_tasks 워커 등록
- vendor/rag synonyms.json 사전 갱신 (RAG 인수자 자산)
- 통합/단위 테스트 다수 추가

+++ api/src/schemas/me.py
@@ -13,7 +13,7 @@ class QuotaResponse(BaseModel):
     reset_at: str
     show_upgrade_prompt: bool
     show_subscribe_button: bool
-    delay_seconds: int
+    delay_seconds: float
 
 
 class UsageSummaryResponse(BaseModel):
